SVGをkonvaに移植した
