+++ frontend/package.json
@@ -3,10 +3,13 @@
     "@types/node": "^20.3.2",
     "@types/react": "^18.2.14",
     "@types/react-dom": "^18.2.6",
+    "canvas": "^2.11.2",
+    "konva": "^9.2.0",
     "next": "^13.4.7",
     "prettier": "^2.8.8",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-konva": "^18.2.10",
     "typescript": "^5.1.6"
   },
   "scripts": {
